requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
tqdm
requests
umap-learn
scipy
